backend/requirements.txt
# Autonomous Diligence Cloud - Python Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Environment & Configuration
python-dotenv>=1.0.0

# AI & LLM
openai>=1.3.0
requests>=2.31.0

# Vector Database
chromadb>=0.4.18

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
pandas>=2.2.0

# Data Models
pydantic>=2.5.0

# Observability & Monitoring
arize-phoenix>=4.0.0
openinference-instrumentation-openai>=0.1.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0

# Optional: For advanced features
# sentence-transformers>=2.2.2  # Uncomment for local embeddings
# pytesseract>=0.3.10  # Uncomment for OCR support

